+++ packages/waku/src/lib/middleware/rsc/ssr.ts
@@ -1,20 +1,107 @@
 import path from 'node:path'; // TODO no node dependency
 import url from 'node:url'; // TODO no node dependency
 
-import { createElement } from 'react';
 import type { ReactNode, FunctionComponent, ComponentProps } from 'react';
-import RDServer from 'react-dom/server.edge';
-import RSDWClient from 'react-server-dom-webpack/client.edge';
 import type { ViteDevServer } from 'vite';
 
 import { resolveConfig, viteInlineConfig } from '../../config.js';
 import { defineEntries } from '../../../server.js';
-import { ServerRoot } from '../../../client.js';
 import { renderRSC } from './worker-api.js';
 import { hasStatusCode, concatUint8Arrays, normalizePath } from './utils.js';
 
-const { renderToReadableStream } = RDServer;
-const { createFromReadableStream } = RSDWClient;
+const loadReact = async (
+  config: Awaited<ReturnType<typeof resolveConfig>>,
+  command: 'dev' | 'build' | 'start',
+) => {
+  if (command !== 'dev') {
+    return (
+      await import(
+        url
+          .pathToFileURL(
+            path.join(
+              config.rootDir,
+              config.distDir,
+              config.publicDir,
+              'assets',
+              'react.js',
+            ),
+          )
+          .toString()
+      )
+    ).default;
+  }
+  return import('react');
+};
+
+const loadRDServer = async (
+  config: Awaited<ReturnType<typeof resolveConfig>>,
+  command: 'dev' | 'build' | 'start',
+) => {
+  if (command !== 'dev') {
+    return (
+      await import(
+        url
+          .pathToFileURL(
+            path.join(
+              config.rootDir,
+              config.distDir,
+              config.publicDir,
+              'assets',
+              'rd-server.js',
+            ),
+          )
+          .toString()
+      )
+    ).default;
+  }
+  return import('react-dom/server.edge');
+};
+
+const loadRSDWClient = async (
+  config: Awaited<ReturnType<typeof resolveConfig>>,
+  command: 'dev' | 'build' | 'start',
+) => {
+  if (command !== 'dev') {
+    return (
+      await import(
+        url
+          .pathToFileURL(
+            path.join(
+              config.rootDir,
+              config.distDir,
+              config.publicDir,
+              'assets',
+              'rsdw-client.js',
+            ),
+          )
+          .toString()
+      )
+    ).default;
+  }
+  return import('react-server-dom-webpack/client.edge');
+};
+
+const loadWakuClient = async (
+  config: Awaited<ReturnType<typeof resolveConfig>>,
+  command: 'dev' | 'build' | 'start',
+) => {
+  if (command !== 'dev') {
+    return import(
+      url
+        .pathToFileURL(
+          path.join(
+            config.rootDir,
+            config.distDir,
+            config.publicDir,
+            'assets',
+            'waku-client.js',
+          ),
+        )
+        .toString()
+    );
+  }
+  return import('waku/client');
+};
 
 // HACK for react-server-dom-webpack without webpack
 const moduleCache = new Map();
@@ -27,10 +114,6 @@ const moduleCache = new Map();
 
 type Entries = {
   default: ReturnType<typeof defineEntries>;
-  resolveClientPath?: (
-    filePath: string,
-    invert?: boolean,
-  ) => string | undefined;
 };
 
 let lastViteServer: ViteDevServer | undefined;
@@ -249,10 +332,20 @@ export const renderHtml = async <Context>(
   htmlStr: string, // Hope stream works, but it'd be too tricky
   context: Context,
 ): Promise<readonly [ReadableStream, Context] | null> => {
+  const [
+    { createElement },
+    { renderToReadableStream },
+    { createFromReadableStream },
+    { ServerRoot, Slot },
+  ] = await Promise.all([
+    loadReact(config, command),
+    loadRDServer(config, command),
+    loadRSDWClient(config, command),
+    loadWakuClient(config, command),
+  ]);
   const entriesFile = getEntriesFile(config, command);
   const {
     default: { getSsrConfig },
-    resolveClientPath,
   } = await (loadServerFile(entriesFile, command) as Promise<Entries>);
   const ssrConfig = await getSsrConfig?.(pathStr);
   if (!ssrConfig) {
@@ -326,19 +419,15 @@ export const renderHtml = async <Context>(
                 return { id, chunks: [id], name };
               }
               // command !== 'dev'
-              const origFile = resolveClientPath?.(
-                path.join(config.rootDir, config.distDir, file),
-                true,
-              );
-              if (
-                origFile &&
-                !origFile.startsWith(path.join(config.rootDir, config.srcDir))
-              ) {
-                const id = url.pathToFileURL(origFile).toString();
-                return { id, chunks: [id], name };
-              }
               const id = url
-                .pathToFileURL(path.join(config.rootDir, config.distDir, file))
+                .pathToFileURL(
+                  path.join(
+                    config.rootDir,
+                    config.distDir,
+                    config.publicDir,
+                    file,
+                  ),
+                )
                 .toString();
               return { id, chunks: [id], name };
             },
@@ -348,9 +437,12 @@ export const renderHtml = async <Context>(
     },
   );
   const [copied, interleave] = injectRscPayload(stream, ssrConfig.input);
-  const elements = createFromReadableStream<Record<string, ReactNode>>(copied, {
-    ssrManifest: { moduleMap, moduleLoading: null },
-  });
+  const elements: Promise<Record<string, ReactNode>> = createFromReadableStream(
+    copied,
+    {
+      ssrManifest: { moduleMap, moduleLoading: null },
+    },
+  );
   const readable = (
     await renderToReadableStream(
       createElement(
@@ -358,8 +450,13 @@ export const renderHtml = async <Context>(
           Omit<ComponentProps<typeof ServerRoot>, 'children'>
         >,
         { elements },
-        ssrConfig.unstable_render(),
+        ssrConfig.unstable_render({ createElement, Slot }),
       ),
+      {
+        onError(err: unknown) {
+          console.error(err);
+        },
+      },
     )
   )
     .pipeThrough(rectifyHtml())