+++ packages/waku/src/lib/middleware/rsc/worker-api.ts
@@ -58,11 +58,16 @@ const getWorker = (command: 'dev' | 'build' | 'start') => {
   const IS_NODE_18 = Number(process.versions.node.split('.')[0]) < 20;
   const worker = new Worker(new URL('worker-impl.js', import.meta.url), {
     env: { __WAKU_CWD__: getCwd() },
-    execArgv: [
-      ...(IS_NODE_18 ? ['--experimental-loader', 'waku/node-loader'] : []),
-      '--conditions',
-      'react-server',
-    ],
+    execArgv:
+      command !== 'dev'
+        ? []
+        : [
+            ...(IS_NODE_18
+              ? ['--experimental-loader', 'waku/node-loader']
+              : []),
+            '--conditions',
+            'react-server',
+          ],
   });
   worker.on('message', (mesg: MessageRes) => {
     if ('id' in mesg) {